add new project changes
