+++ src/main/java/ru/otus/library/authentication/SecurityConfig.java
@@ -88,17 +88,12 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers(HttpMethod.DELETE,"/books")
                 .hasRole("ADMIN")
                 .and()
-                .authorizeRequests()
-                .antMatchers(HttpMethod.POST, "/perform_logout")
-                .permitAll()
-                .and()
                 .formLogin()
-                .loginProcessingUrl("/perform_login")
                 .successHandler(successHandler)
                 .failureHandler(failureHandler)
                 .usernameParameter("username")
                 .passwordParameter("password")
-                .defaultSuccessUrl("/library")
+                .defaultSuccessUrl("/")
                 .and()
                 .rememberMe().key("someSecret").alwaysRemember(true)
                 .and()
