+++ src/main/java/ru/otus/library/dto/ShowStatisticsDto.java
@@ -1,0 +1,17 @@
+package ru.otus.library.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class ShowStatisticsDto {
+
+    private String title;
+
+    private String userName;
+
+    private String readDate;
+}
