+++ src/main/java/ru/otus/library/service/StatisticsService.java
@@ -1,8 +1,13 @@
 package ru.otus.library.service;
 
 import ru.otus.library.dto.StatisticsDto;
+import ru.otus.library.model.entity.ReadStatistics;
+
+import java.util.List;
 
 public interface StatisticsService {
+
+    List<ReadStatistics> findAllStatistics();
 
     StatisticsDto findStatisticsByUserAndBook(final String username, final String title);
