+++ src/ui/package.json
@@ -23,7 +23,7 @@
     "react-bootstrap": "^2.7.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.8.1",
-    "react-router-dom": "^6.6.1",
+    "react-router-dom": "^6.8.2",
     "react-scripts": "^5.0.1",
     "react-stomp": "^5.1.0",
     "sass-loader": "^13.2.0",
