+++ src/ui/src/styles/statistics.module.scss
@@ -1,0 +1,55 @@
+@font-face {
+  font-family: "SBSansText";
+  src: url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.eot");
+  src: url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.eot?#iefix") format("embedded-opentype"),
+  url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.woff2") format("woff2"),
+  url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.woff") format("woff"),
+  url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.ttf") format("truetype"),
+  url("//db.onlinewebfonts.com/t/c68e21ad1a10803b1ef491fd0392b067.svg#SB Sans Text") format("svg");
+}
+
+.link {
+  font-size: 20px;
+  margin-right: 25px;
+  cursor: pointer;
+  font-weight: bold;
+}
+
+.link:hover {
+  color: #3f3479;
+  transition: all 0.3s ease 0s;
+}
+
+.stat_container {
+  overflow: hidden;
+  background-color: white;
+  width: 570px;
+  height: 300px;
+  overflow-y: auto;
+  padding: 20px;
+}
+
+.table_header {
+
+}
+
+.tr_header {
+
+}
+
+.td_header {
+  padding: 10px;
+  min-width: 180px;
+  background: #5e7967;
+  color: white;
+}
+
+.tr {
+
+}
+
+.td {
+  padding: 10px;
+  min-width: 180px;
+  background: #dee2df;
+}
